pyproject.toml
[project]
name = "study-rag-agent"
version = "0.1.0"
description = "StudyRAG - Agent RAG local avec Docling et Ollama pour étudiants"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "pydantic-ai>=0.7.4",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",
    "numpy>=2.0.2",
    "ollama>=0.4.0",
    "docling[vlm]>=2.55.0",
    "chromadb>=0.5.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "sentence-transformers>=3.0.0",
    "openai>=1.0.0",
    "structlog>=25.5.0",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
